requirements.txt
torch>=1.11.0
ollama>=0.4.7
pymupdf>=1.18.19
transformers>=4.34.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
flask>=2.0.0
flask-restful>=0.3.9
flasgger>=0.9.5
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
##psycopg2>=2.9.6##
vosk==0.3.45
telebot
requests==2.32.3
reportlab==4.3.1
